speech/Porcupine/requirements.txt
pvporcupine==3.0.2
pvrecorder==1.2.2
python-dotenv==1.0.1